+++ src/HEADERS/bar_ops_SSE.h
@@ -6,7 +6,7 @@
 #define BAR_OPS_SSE_H
 
 /**
-   @fn double complex baryon_contract( const struct spinor DiQ , const struct spinor S , const int d0 , const int d1 , const int d2 , const int d3 )
+   @fn double complex baryon_contract( const struct spinor DiQ , const struct spinor S , const size_t d0 , const size_t d1 , const size_t d2 , const size_t d3 )
    @brief color trace a diquark with a propagator
    @return the color trace
  */
@@ -14,10 +14,10 @@
 double complex
 baryon_contract( const struct spinor DiQ ,
 		 const struct spinor S ,
-		 const int d0 ,
-		 const int d1 ,
-		 const int d2 ,
-		 const int d3 ) ;
+		 const size_t d0 ,
+		 const size_t d1 ,
+		 const size_t d2 ,
+		 const size_t d3 ) ;
 
 /**
    @fn void cross_color_trace( struct spinor *__restrict DiQ , const struct spinor S )
