+++ src/HEADERS/currents.h
@@ -6,7 +6,7 @@
 #define CURRENTS_H
 
 /**
-   @fn void contract_conserved_local_site( struct PIdata *DATA_AA , struct PIdata *DATA_VV ,  const struct site *lat , const struct spinor *S1 , const struct spinor *S1UP , const struct spinor *S2 , const struct spinor *S2UP , const struct gamma *GAMMAS , const int AGMAP[ ND ] , const int VGMAP[ ND ] , const int x , const int t ) 
+   @fn void contract_conserved_local_site( struct PIdata *DATA_AA , struct PIdata *DATA_VV ,  const struct site *lat , const struct spinor *S1 , const struct spinor *S1UP , const struct spinor *S2 , const struct spinor *S2UP , const struct gamma *GAMMAS , const size_t AGMAP[ ND ] , const size_t VGMAP[ ND ] , const size_t x , const size_t t ) 
    @brief conserved-local Wilson current at a single site x + LCU * t
  */
 void
@@ -18,14 +18,14 @@ contract_conserved_local_site( struct PIdata *DATA_AA ,
 			       const struct spinor *S2 ,
 			       const struct spinor *S2UP ,
 			       const struct gamma *GAMMAS ,
-			       const int AGMAP[ ND ] ,
-			       const int VGMAP[ ND ] ,
-			       const int x ,
-			       const int t ) ;
+			       const size_t AGMAP[ ND ] ,
+			       const size_t VGMAP[ ND ] ,
+			       const size_t x ,
+			       const size_t t ) ;
 
 
 /**
-   @fn void contract_local_local( struct PIdata *DATA_AA , struct PIdata *DATA_VV , const struct spinor *S1 , const struct spinor *S2 , const struct gamma *GAMMAS , const int AGMAP[ ND ] , const int VGMAP[ ND ] , const int x , const int t ) 
+   @fn void contract_local_local( struct PIdata *DATA_AA , struct PIdata *DATA_VV , const struct spinor *S1 , const struct spinor *S2 , const struct gamma *GAMMAS , const size_t AGMAP[ ND ] , const size_t VGMAP[ ND ] , const size_t x , const size_t t ) 
    @brief local-local vector and axial currents
 */
 void
@@ -34,9 +34,9 @@ contract_local_local_site( struct PIdata *DATA_AA ,
 			   const struct spinor *S1 ,
 			   const struct spinor *S2 ,
 			   const struct gamma *GAMMAS ,
-			   const int AGMAP[ ND ] ,
-			   const int VGMAP[ ND ] ,
-			   const int x ,
-			   const int t ) ;
+			   const size_t AGMAP[ ND ] ,
+			   const size_t VGMAP[ ND ] ,
+			   const size_t x ,
+			   const size_t t ) ;
 
 #endif
