+++ src/VPF/cl_offdiagonal.c
@@ -25,10 +25,10 @@ cl_offdiagonal( struct propagator prop1 ,
 		const char *outfile )
 {
   // vector gamma map -> x,y,z,t
-  const int VGMAP[ ND ] = { GAMMA_0 , GAMMA_1 , GAMMA_2 , GAMMA_3 } ;
+  const size_t VGMAP[ ND ] = { GAMMA_0 , GAMMA_1 , GAMMA_2 , GAMMA_3 } ;
 
   // need to look these up
-  const int AGMAP[ ND ] = { GAMMA_5 + 1 , GAMMA_5 + 2 , GAMMA_5 + 3 , GAMMA_5 + 4 } ;
+  const size_t AGMAP[ ND ] = { GAMMA_5 + 1 , GAMMA_5 + 2 , GAMMA_5 + 3 , GAMMA_5 + 4 } ;
 
   // spinor storage
   struct spinor *S1 = NULL , *S1f = NULL , *S1END = NULL , *S1UP = NULL ;
@@ -97,7 +97,7 @@ cl_offdiagonal( struct propagator prop1 ,
   }
 
   // copy for the final timeslice
-  int x ;
+  size_t x ;
 #pragma omp parallel for private(x)
   for( x = 0 ; x < LCU ; x++ ) {
     equate_spinor_minus( &S1END[x] , &S1[x] ) ;
@@ -105,7 +105,7 @@ cl_offdiagonal( struct propagator prop1 ,
   }
 
   // loop the timeslices
-  int t ;
+  size_t t ;
   for( t = 0 ; t < LT-1 ; t++ ) {
 
     // multiple time source support
